lifter: Extract some shared lifter state to Exalt.[hpp,cpp].
